// ReactJs/homework-3-class-05/fake-store-api-e-comerce-app/src/types/Product.ts
export interface Product {
    id: number;
    title: string;
    price: number;
    description: string;
    category: string;
    image: string;
    rating: {
        rate: number;
        count: number;
    };
}

export interface ApiError {
    message: string;
    status?: number;
    data? : any;
}